+++ src/Components/FundsRequired.js
@@ -15,22 +15,23 @@ function FundsRequired({ address }) {
         To be able to interract fully with the Dash Platform and create a
         username you'll need some Dash in your wallet.
       </p>
-      <TextDropdown title="Get some Dash instantly">
-        <OrderedList>
-          <li>
-            <a
-              href="http://faucet.evonet.networks.dash.org"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              Go to the Dash Faucet here
-            </a>
-          </li>
-          <li>Copy your wallet address below and paste into the faucet</li>
-          <li>Fill in the verification letters on the Faucet</li>
-          <li>Click "Get coins". Your balance will update below.</li>
-        </OrderedList>
-      </TextDropdown>
+      <h3>Get Dash instantly</h3>
+      <OrderedList>
+        <li>
+          <a
+            href="http://faucet.evonet.networks.dash.org"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            Go to the Dash Faucet here
+          </a>
+        </li>
+        <li>
+          Follow the instructions on the Faucet, inputting your Dash Address
+        </li>
+        <li>Your balance will update below</li>
+      </OrderedList>
+
       <h3>Your Dash Address</h3>
       <DashAddress>{address}</DashAddress>
       <AwaitPayment />